// src/syndicate/enums/storage.ts
export enum SessionStorage {
  SYNDICATE_SORTING_DEFAULTS = 'syndicate-sorting-defaults',
  SYNDICATE_SORTING_ALARMS = 'syndicate-sorting-alarms',
  SYNDICATE_SORTING_HISTORICAL = 'syndicate-sorting-historical',
  SYNDICATE_SORTING_REPORTS = 'syndicate-sorting-reports',
  SYNDICATE_SORTING_VARIABLES = 'syndicate-sorting-variables',
  SYNDICATE_SORTING_VARIABLES_CONFIG = 'syndicate-sorting-variables-config',
  SYNDICATE_SORTING_USERS = 'syndicate-sorting-users',
  SYNDICATE_SORTING_PROFILES = 'syndicate-sorting-profiles',
  SYNDICATE_SORTING_VIEWS = 'syndicate-sorting-views',
  SYNDICATE_VARIABLES = 'syndicate-variables',
  SYNDICATE_VARIABLES_CATEGORY_SEARCH = 'syndicate-variables-category-search',
  SYNDICATE_SORTING_CONDITIONS = 'syndicate-sorting-conditions',
  SYNDICATE_GRAPH_SCALING = 'syndicate-graph-scaling',
  SYNDICATE_VARIABLES_DATES = 'syndicate-variables-dates',
  SYNDICATE_GRAPH_SCALING_VARIABLE = 'syndicate-graph-scaling-variable',
  SYNDICATE_SORTING_DOCUMENTS = 'syndicate-sorting-documents',
  SYNDICATE_SORTING_STATIONS = 'syndicate-sorting-stations',
  SYNDICATE_SORTING_DEVICES = 'syndicate-sorting-devices',
  SYNDICATE_SORTING_COUNTERS = 'syndicate-sorting-counters',
  SYNDICATE_SORTING_COUNTERS_CONFIG = 'syndicate-sorting-counters-config'
}

export enum LocalStorage {
  SYNDICATE_CATEGORIES = 'syndicate-categories',
  SYNDICATE_TYPES = 'syndicate-types',
  SYNDICATE_TYPES_DOCUMENT = 'syndicate-types-document',
  SYNDICATE_CATEGORIES_DOCUMENT = 'syndicate-categories-document'
}